+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `note`:
+  - `RandomSeed` and its functions `from_bytes` and `as_bytes` now public
+  -  `Note::from_parts` now public
+- `keys::Diversifier::from_bytes` now public
 
 ## [0.2.0] - 2022-06-24
 ### Added